+++ packages/web/src/alpha/select/__stories__/MultiSelect.stories.tsx
@@ -4,6 +4,7 @@ import { useMultiSelect } from '@coinbase/cds-common/select/useMultiSelect';
 import { css } from '@linaria/core';
 
 import { Icon } from '../../../icons/Icon';
+import { VStack } from '../../../layout/VStack';
 import { Pressable } from '../../../system/Pressable';
 import { Text } from '../../../typography/Text';
 import { Select, type SelectOptionComponent } from '../Select';
@@ -166,6 +167,83 @@ export const HideSelectAll = () => {
       type="multi"
       value={value}
     />
+  );
+};
+
+export const Alignments = () => {
+  const exampleOptions = [
+    { value: null, label: 'Remove selection' },
+    { value: '1', label: 'Option 1' },
+    { value: '2', label: 'Option 2' },
+    { value: '3', label: 'Option 3' },
+    { value: '4', label: 'Option 4' },
+    { value: '5', label: 'Option 5' },
+    { value: '6', label: 'Option 6' },
+    { value: '7', label: 'Option 7' },
+    { value: '8', label: 'Option 8' },
+  ];
+  const { value, onChange } = useMultiSelect({
+    initialValue: ['1'],
+  });
+
+  return (
+    <VStack gap={2}>
+      <Select
+        label="Default align - start"
+        onChange={onChange}
+        options={exampleOptions}
+        placeholder="Empty value"
+        type="multi"
+        value={value}
+      />
+      <Select
+        align="center"
+        label="Center align"
+        onChange={onChange}
+        options={exampleOptions}
+        placeholder="Empty value"
+        type="multi"
+        value={value}
+      />
+      <Select
+        align="end"
+        label="End align"
+        onChange={onChange}
+        options={exampleOptions}
+        placeholder="Empty value"
+        type="multi"
+        value={value}
+      />
+      <Select
+        compact
+        label="Compact align - start"
+        onChange={onChange}
+        options={exampleOptions}
+        placeholder="Empty value"
+        type="multi"
+        value={value}
+      />
+      <Select
+        compact
+        align="center"
+        label="Compact align - center"
+        onChange={onChange}
+        options={exampleOptions}
+        placeholder="Empty value"
+        type="multi"
+        value={value}
+      />
+      <Select
+        compact
+        align="end"
+        label="Compact align - end"
+        onChange={onChange}
+        options={exampleOptions}
+        placeholder="Empty value"
+        type="multi"
+        value={value}
+      />
+    </VStack>
   );
 };