+++ packages/web/CHANGELOG.md
@@ -8,11 +8,17 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
-## 8.41.0 (2/4/2026 PST)
+## 8.42.0 (2/4/2026 PST)
 
 #### 🚀 Updates
 
 - Add new tray design. [[#349](https://github.com/coinbase/cds/pull/349)]
+
+## 8.41.0 (2/4/2026 PST)
+
+#### 🚀 Updates
+
+- Add align prop to Select and Combobox. [[#348](https://github.com/coinbase/cds/pull/348)]
 
 ## 8.40.2 ((2/2/2026, 11:25 AM PST))